move drawTree method to SyntaxNode
